+++ fx-util/src/main/java/com/dua3/fx/util/FxUtil.java
@@ -3,9 +3,66 @@ package com.dua3.fx.util;
 import javafx.geometry.Bounds;
 import javafx.geometry.Dimension2D;
 import javafx.scene.text.Font;
+import javafx.scene.text.FontPosture;
+import javafx.scene.text.FontWeight;
 import javafx.scene.text.Text;
 
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+
 public final class FxUtil {
+
+    private static class FontHelper {
+        private static final List<String> FONTS_ALL;
+        private static final List<String> FONTS_PROPORTIONAL;
+        private static final List<String> FONTS_MONOSPACED;
+
+        static {
+            List<String> all = Font.getFamilies();
+            all.sort(String::compareToIgnoreCase);
+
+            List<String> proportional = new ArrayList<>();
+            List<String> monospaced = new ArrayList<>();
+
+            Text thin = new Text("1 l");
+            Text thick = new Text("M_W");
+
+            for (String family : all) {
+                Font font = Font.font(family, FontWeight.NORMAL, FontPosture.REGULAR, 14.0d);
+                thin.setFont(font);
+                thick.setFont(font);
+                if (thin.getLayoutBounds().getWidth() == thick.getLayoutBounds().getWidth()) {
+                    monospaced.add(family);
+                } else {
+                    proportional.add(family);
+                }
+            }
+
+            FONTS_ALL = Collections.unmodifiableList(all);
+            FONTS_PROPORTIONAL = Collections.unmodifiableList(proportional);
+            FONTS_MONOSPACED = Collections.unmodifiableList(monospaced);
+        }
+    }
+
+    public enum FontType {
+        PROPORTIONAL,
+        MONOSPACE,
+        ANY
+    }
+    
+    public static List<String> getFonts(FontType t) {
+        switch (t) {
+            case ANY:
+                return FontHelper.FONTS_ALL;
+            case PROPORTIONAL:
+                return FontHelper.FONTS_PROPORTIONAL;
+            case MONOSPACE:
+                return FontHelper.FONTS_MONOSPACED;
+            default:
+                throw new IllegalArgumentException(String.valueOf(t));
+        }
+    }
 
     public static Font convert(com.dua3.utility.text.Font font) {
         return new Font(font.getFamily(), font.getSizeInPoints());
@@ -32,6 +89,6 @@ public final class FxUtil {
     public static Dimension2D growToFit(Dimension2D a, Bounds b) {
         return new Dimension2D(Math.max(a.getWidth(), b.getWidth()), Math.max(a.getHeight(), b.getHeight()));
     }
-    
+
     private FxUtil() {}
 }